refactor: extract cost and shared-knowledge mixins from mem0 adapter
